Move the AssemblyResolve callbacks into another class that finishes its static constructor calls so it is not under a lock, removing a deadlock scenario even if the SDK initialization creates additional Threads/Tasks that trigger instrumentation
